+++ src/index.ts
@@ -1,6 +1,6 @@
-export { pickObjectProperties } from './pickObjectProperties';
-export { omitObjectProperties } from './omitObjectProperties';
-export { mergeObjectProperties } from './mergeObjectProperties';
-export { closeObjectDeep } from './closeObjectDeep';
-export { openObjectDeep } from './openObjectDeep';
-export { requireObjectProperties } from './requireObjectProperties';
+export { pickProps } from './pickProps';
+export { omitProps } from './omitProps';
+export { mergeProps } from './mergeProps';
+export { sealSchema } from './sealSchema';
+export { unsealSchema } from './unsealSchema';
+export { requireProps } from './requireProps';